move code out of main.py
